requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
Django==5.1.2
django-cloudinary-storage==0.3.0
django-countries==7.6.1
django-filter==24.3
djangorestframework==3.15.2
idna==3.10
pillow==11.0.0
pytz==2024.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
